requirements.txt
openai
google-generativeai
pandas
numpy
matplotlib
tensorflow
scikit-learn
typing-extensions